requirements.txt
MarkupSafe==2.1.1
Flask==2.1.0
gunicorn==20.1.0
psycopg[binary]
psycopg[pool]
bcrypt==4.0.1
Flask-Sockets
gevent
pytest